ci: fix cherry-pick job by using newer git
https://github.com/git/git/commit/37897bfc27d3003dddf7722113e676e925fe9e21 made it possible to use
the -m flag with cherry-pick, even when the target is not a merge commit.

This commit changes the image used to run the cherry-pick job to alpine so that we get
a more recent version of git.

The alpine image will also download much faster when the CI node does not have the image cached.